Add `commons-beanutils` v1.9.2 as dependency
`commons-beanutils` is a dependency of `commons-configuration`, and it
seems to be used when `CompilationTaskBuilder#compileProperties()` is
called.

Therefore, this dependency was added so that clients of this library
which use Maven-style transitive dependencies will not need to manually
add `commons-beanutils` as a dependency.

(I'm not sure why `commons-beanutils` isn't already a dependency of
`commons-configuration`. Oh well...)